+++ game/player.py
@@ -7,5 +7,15 @@ class Player:
         self.tiles += BagTiles.take(
             7 - len(BagTiles)
         );
-    def has_letters(self):
-        print(self.tiles)
+    def has_letters(self, tiles):
+        res = []
+        bagtiles = []
+        for _ in range(0, len(tiles)):
+            res += tiles[_].letter
+        for _ in range(0, len(self.tiles)):
+            bagtiles += self.tiles[_].letter
+        for _ in res:
+            if _ in bagtiles:
+                return True;
+            else:
+                return False;